+++ test/model.js
@@ -49,6 +49,7 @@ metatests.test('Model: from struct', (test) => {
   test.strictEqual(model.entities.get('Company'), {
     name: 'Company',
     namespaces: new Set(),
+    parent: '',
     kind: 'dictionary',
     scope: 'application',
     store: 'persistent',
@@ -75,10 +76,10 @@ metatests.test('Model: from struct', (test) => {
 
 metatests.test('Model: loader', async (test) => {
   const model = await Model.load(process.cwd() + '/test/schemas', types);
-  test.strictEqual(model.entities.size, 4);
+  test.strictEqual(model.entities.size, 5);
   const Account = model.entities.get('Account');
   test.strictEqual(Account.fields.fullName.constructor.name, 'Schema');
-  test.strictEqual(model.order.size, 4);
+  test.strictEqual(model.order.size, 5);
   test.strictEqual(typeof model.types, 'object');
   test.strictEqual(typeof model.database, 'object');
   test.end();